+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 ## 3.2.0 | 2019.11.29
-- 
+- Won't prompt to reload after extension update when using default config
+
+- Remove update notification
 
 ## 3.1.2 | 2019.11.25
 - Fixed template expression highlight error [#358](https://github.com/Binaryify/OneDark-Pro/issues/358)